--- crates/test-spidermonkey/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "test-spidermonkey"
-version = "0.1.0"
-edition = "2018"
-publish = false
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.40"
-env_logger = "0.8.4"
-log = "0.4.14"
-wasmparser = "0.78.2"
-wasmtime = "0.30.0"
-wasmtime-wasi = "0.30.0"
-witx-bindgen-wasmtime = { path = "../wasmtime" }
-witx-bindgen-gen-core = { path = "../gen-core" }
-witx-bindgen-gen-spidermonkey = { path = "../gen-spidermonkey" }
-lazy_static = "1.4.0"
