--- crates/test-spidermonkey/tests/all/main.rs
@@ -1,3 +1,0 @@
-mod functions;
-mod lists;
-mod strings;
