+++ tests/runtime/smw_functions/exports.witx
@@ -1,0 +1,10 @@
+test_imports: function()
+
+f1: function()
+f2: function(a: u32)
+f3: function(a: u32, b: u32)
+
+f4: function() -> u32
+f5: function() -> (u32, u32)
+
+f6: function(a: u32, b: u32, c: u32) -> (u32, u32, u32)
