+++ tests/runtime/smw_functions/wasm.js
@@ -1,3 +1,5 @@
+import * as imports from "imports";
+
 function assert(condition, message) {
   if (!condition) {
     throw new Error(message);
@@ -6,6 +8,43 @@ function assert(condition, message) {
 
 function assertEq(a, b) {
   assert(a == b, `assertEq failed: ${a} != ${b}`);
+}
+
+export function test_imports() {
+  const { f1, f2, f3, f4, f5, f6 } = imports;
+
+  //
+  // Testing arguments.
+  //
+
+  f1();
+
+  f2(42);
+
+  // Min and max `u32`.
+  f3(0, 4294967295);
+
+  //
+  // Testing returns.
+  //
+
+  {
+    const a = f4();
+    assertEq(a, 1337);
+  }
+
+  {
+    const [a, b] = f5();
+    assertEq(a, 1);
+    assertEq(b, 2);
+  }
+
+  {
+    const [a, b, c] = f6(100, 200, 300);
+    assertEq(a, 101);
+    assertEq(b, 201);
+    assertEq(c, 301);
+  }
 }
 
 //
