+++ tests/runtime/smw_lists/exports.witx
@@ -1,0 +1,6 @@
+test_imports: function()
+
+f1: function(l: list<u32>)
+f2: function() -> list<u32>
+f3: function(a: list<u32>, b: list<u32>) -> (list<u32>, list<u32>)
+f4: function(l: list<list<u32>>) -> list<list<u32>>
