+++ tests/runtime/smw_strings/exports.witx
@@ -1,0 +1,5 @@
+test_imports: function()
+
+f1: function(s: string)
+f2: function() -> string
+f3: function(a: string, b:string, c: string) -> (string, string, string)
